Update GameBoard.html
So, this is crude. I need to practice my JS more it seems. Probably not a good implementation of this idea, but the idea is basically that you could have mouse-over interactivity. Obviously the X would be centered and resized. I suppose I could have used an image file for the X instead of text, but I just wanted to get a bare bones prototype up here.

In either case, whether we use mouse-over interaction or input based interaction, I'm pretty sure the next step would be implementing some kind of validation check to see if the box you're trying to mark is open or not, and displaying some kind of feedback. 

Moving forward in this type of interaction, the next step (besides making it look better) would be to allow the user to click and mark a box (same with text input, input a box and it gets marked on screen).